restrict updating to owner

+++ src/domain/companies/companies.controller.ts
@@ -45,9 +45,11 @@ class Controller {
       address,
       location,
     } = req.body;
+    const { id: userId } = req.user!;
 
     const { data } = await Service.update({
       id,
+      userId,
       name,
       phone,
       email,
@@ -63,8 +65,9 @@ class Controller {
 
   destroy = async (req: Request, res: Response) => {
     const { id } = req.params;
+    const { id: userId } = req.user!;
 
-    await Service.destroy({ id });
+    await Service.destroy({ id, userId });
 
     res.status(StatusCode.NoContent).json({});
   };
